requirements.txt
python-dotenv
requests
openai-agents 
openai-agents[voice]
openai
numpy
sounddevice 
windows-curses